+++ utmap-client/src/components/pages/CalendarPage.js
@@ -1,20 +1,32 @@
-import React, { Component } from "react";
-//import React from 'react'
-import BigCalendar from '../BigCalendar'
-//import ReactDOM from "react-dom"
+import React, { Component } from 'react'
+import { Calendar, momentLocalizer } from "react-big-calendar";
+import moment from 'moment';
+import "react-big-calendar/lib/css/react-big-calendar.css";
+
+moment.locale("en-CA");
+const localizer = momentLocalizer(moment);
 
 class CalendarPage extends Component {
+    getStyle = () => {
+        return {
+            height: 700, //needs a strict height or flexbox to show month calendar
+            padding: '20px',
+        }
+    }
     render() {
-        //const params = ReactDOM.findDOMNode(this.props.events);
-        //const { events, match: {props} } = this.events;
-        //console.log(params)
         return (
-            //<p>Calendar</p>
-            <div className="container">
-                <BigCalendar events={this.props.events} />
+            <div style={this.getStyle()}>
+                <Calendar 
+                    localizer={localizer} 
+                    defaultView='month'
+                    step={60}
+                    events={this.props.events} 
+                    startAccessor="start" 
+                    endAccessor="end" 
+                    showMultiDayTimes  
+                />
             </div>
-           
-        )
+        );
     }
 }
